# FatecAppBackend/FatecAppBackend.API/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "FatecApp API",
    "description": "Documentation of Api's endpoints",
    "termsOfService": "https://example.com/terms",
    "contact": {
      "name": "João Vitor",
      "url": "https://github.com/JoVi0li",
      "email": "jovioli.dev04@gmail.com"
    },
    "license": {
      "name": "Nothing",
      "url": "https://example.com/license"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/college/create": {
      "post": {
        "tags": [
          "College"
        ],
        "summary": "Create College",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "College props",
            "schema": {
              "$ref": "#/definitions/FatecAppBackend.Domain.Commands.College.CreateCollegeCommand"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/FatecAppBackend.Shared.Commands.GenericCommandsResult"
            }
          }
        }
      }
    },
    "/api/v1/college/delete/{id}": {
      "delete": {
        "tags": [
          "College"
        ],
        "summary": "Delete College",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "College id",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/FatecAppBackend.Shared.Commands.GenericCommandsResult"
            }
          }
        }
      }
    },
    "/api/v1/college/update": {
      "patch": {
        "tags": [
          "College"
        ],
        "summary": "Update College",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "College props",
            "schema": {
              "$ref": "#/definitions/FatecAppBackend.Domain.Commands.College.UpdateCollegeCommand"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/FatecAppBackend.Shared.Commands.GenericCommandsResult"
            }
          }
        }
      }
    },
...
    "/api/v1/college/id/{id}": {
      "get": {
        "tags": [
          "College"
        ],
        "summary": "Get College by id",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "College props",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/FatecAppBackend.Shared.Queries.GenericQueryResult"
            }
          }
        }
      }
    },
    "/api/v1/event/create": {
      "post": {
        "tags": [
          "Event"
        ],
        "summary": "Create Event",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Event props",
            "schema": {
              "$ref": "#/definitions/FatecAppBackend.Domain.Commands.Event.CreateEventCommand"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/FatecAppBackend.Shared.Commands.GenericCommandsResult"
            }
          }
        }
      }
    },
    "/api/v1/event/delete/{id}": {
      "delete": {
        "tags": [
          "Event"
        ],
        "summary": "Delete Event",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Event id",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/FatecAppBackend.Shared.Commands.GenericCommandsResult"
            }
          }
        }
      }
    },
    "/api/v1/event/update": {
      "patch": {
        "tags": [
          "Event"
        ],
        "summary": "Update Event",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Event props",
            "schema": {
              "$ref": "#/definitions/FatecAppBackend.Domain.Commands.Event.UpdateEventCommand"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/FatecAppBackend.Shared.Commands.GenericCommandsResult"
            }
          }
        }
      }
    },
...
    "/api/v1/event/id/{id}": {
      "get": {
        "tags": [
          "Event"
        ],
        "summary": "Get Event by id",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Event id",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/FatecAppBackend.Shared.Queries.GenericQueryResult"
            }
          }
        }
      }
    },
    "/api/v1/participant/create": {
      "post": {
        "tags": [
          "Participant"
        ],
        "summary": "Create Participant",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Participant props",
            "schema": {
              "$ref": "#/definitions/FatecAppBackend.Domain.Commands.Participant.CreateParticipantCommand"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/FatecAppBackend.Shared.Commands.GenericCommandsResult"
            }
          }
        }
      }
    },
    "/api/v1/participant/delete/{id}": {
      "delete": {
        "tags": [
          "Participant"
        ],
        "summary": "Delete Participant",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Participant id",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/FatecAppBackend.Shared.Commands.GenericCommandsResult"
            }
          }
        }
      }
    },
    "/api/v1/participant/id/{id}": {
      "get": {
        "tags": [
          "Participant"
        ],
        "summary": "Get Participant by id",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Participant id",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/FatecAppBackend.Shared.Queries.GenericQueryResult"
            }
          }
        }
      }
    },
    "/api/v1/participant/eventid/{eventId}": {
      "get": {
        "tags": [
          "Participant"
        ],
        "summary": "Get Participant by Event id",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "eventId",
            "description": "Event id",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/FatecAppBackend.Shared.Queries.GenericQueryResult"
            }
          }
        }
      }
    },
    "/api/v1/participant/usercollegeid/{userCollegeId}": {
      "get": {
        "tags": [
          "Participant"
        ],
        "summary": "Get Participant by UserCollege id",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userCollegeId",
            "description": "UserCollege id",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/FatecAppBackend.Shared.Queries.GenericQueryResult"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "FatecAppBackend.Domain.Commands.College.CreateCollegeCommand": {
      "type": "object",
      "properties": {
        "notifications": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Flunt.Notifications.Notification"
          },
          "readOnly": true
        },
        "isValid": {
          "type": "boolean",
          "readOnly": true
        },
        "name": {
          "type": "string"
        },
        "course": {
          "type": "string"
        },
        "time": {
          "$ref": "#/definitions/FatecAppBackend.Shared.EnTime"
        },
        "localization": {
          "type": "string"
        }
      }
    },
    "FatecAppBackend.Domain.Commands.College.UpdateCollegeCommand": {
      "type": "object",
      "properties": {
        "notifications": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Flunt.Notifications.Notification"
          },
          "readOnly": true
        },
        "isValid": {
          "type": "boolean",
          "readOnly": true
        },
        "college": {
          "$ref": "#/definitions/FatecAppBackend.Shared.DTOs.College.UpdateCollegeDTO"
        }
      }
    },
    "FatecAppBackend.Domain.Commands.Event.CreateEventCommand": {
      "type": "object",
      "properties": {
        "notifications": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Flunt.Notifications.Notification"
          },
          "readOnly": true
        },
        "isValid": {
          "type": "boolean",
          "readOnly": true
        },
        "eventOwnerId": {
          "format": "uuid",
          "type": "string"
        },
        "from": {
          "type": "string"
        },
        "to": {
          "type": "string"
        },
        "route": {
          "type": "string"
        },
        "onlyWomen": {
          "type": "boolean"
        },
        "timeToGo": {
          "type": "string"
        },
        "status": {
          "$ref": "#/definitions/FatecAppBackend.Shared.Enums.EnStatus"
        }
      }
    },
    "FatecAppBackend.Domain.Commands.Event.UpdateEventCommand": {
      "type": "object",
      "properties": {
        "notifications": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Flunt.Notifications.Notification"
          },
          "readOnly": true
        },
        "isValid": {
          "type": "boolean",
          "readOnly": true
        },
        "event": {
          "$ref": "#/definitions/FatecAppBackend.Shared.DTOs.Event.UpdateEventDTO"
        }
      }
    },
    "FatecAppBackend.Domain.Commands.Participant.CreateParticipantCommand": {
      "type": "object",
      "properties": {
        "notifications": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Flunt.Notifications.Notification"
          },
          "readOnly": true
        },
        "isValid": {
          "type": "boolean",
          "readOnly": true
        },
        "eventId": {
          "format": "uuid",
          "type": "string"
        },
        "userCollegeId": {
          "format": "uuid",
          "type": "string"
        }
      }
    },
...
    "FatecAppBackend.Shared.Commands.GenericCommandsResult": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "message": {
          "type": "string"
        },
        "data": { }
      }
    },
    "FatecAppBackend.Shared.DTOs.College.UpdateCollegeDTO": {
      "type": "object",
      "properties": {
        "id": {
          "format": "uuid",
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "course": {
          "type": "string"
        },
        "time": {
          "$ref": "#/definitions/FatecAppBackend.Shared.EnTime"
        },
        "localization": {
          "type": "string"
        }
      }
    },
    "FatecAppBackend.Shared.DTOs.Event.UpdateEventDTO": {
      "type": "object",
      "properties": {
        "id": {
          "format": "uuid",
          "type": "string"
        },
        "from": {
          "type": "string"
        },
        "to": {
          "type": "string"
        },
        "route": {
          "type": "string"
        },
        "onlyWomen": {
          "type": "boolean"
        },
        "timeToGo": {
          "format": "date-time",
          "type": "string"
        },
        "status": {
          "$ref": "#/definitions/FatecAppBackend.Shared.Enums.EnStatus"
        }
      }
    },
...
    "FatecAppBackend.Shared.EnTime": {
      "format": "int32",
      "enum": [
        0,
        1
      ],
      "type": "integer"
    },
    "FatecAppBackend.Shared.Enums.EnStatus": {
      "format": "int32",
      "enum": [
        0,
        1,
        2,
        3
      ],
      "type": "integer"
    },
    "FatecAppBackend.Shared.Queries.GenericQueryResult": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "message": {
          "type": "string"
        },
        "data": { }
      }
    },
    "Flunt.Notifications.Notification": {
      "type": "object",
      "properties": {
        "key": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    }
  }
}